+++ software_model/utils.h
@@ -20,7 +20,7 @@ void image_free(image_t image);
 
 image_t extract_segment(image_t image, uint32_t start_x, uint32_t start_y, uint16_t rows, uint16_t cols);
 
-double* get_scaled_coordinates(unsigned c, double sc);
+float* get_scaled_coordinates(unsigned c, float sc);
 
 image_t bin2image(const char* filename);
 void save_to_pgm(const char* filename, image_t image);
@@ -28,6 +28,6 @@ void save_to_bin(const char* filename, image_t image);
 
 image_t invert_image(image_t image);
 
-image_t bilinear_scaling(image_t input, double sx, double sy);
+image_t bilinear_scaling(image_t input, float sx, float sy);
 
 #endif
